re-organize structures
