+++ src/spu/src/sc/dispatcher.rs
@@ -31,7 +31,7 @@ use flv_util::actions::Actions;
 
 use crate::core::SharedGlobalContext;
 use crate::core::SpecChange;
-use crate::controllers::leader_replica::{LeaderReplicaState, LeaderReplicaControllerCommand};
+use crate::replication::leader_replica::{LeaderReplicaState, LeaderReplicaControllerCommand};
 use crate::InternalServerError;
 
 use super::SupervisorCommand;
